rename: validation에서 nickname-validation으로 파일명 변경 (#67)
