+++ lib/utils/nickname-validation.ts
